+++ launcher/app/src/main/java/com/blankj/launcher/app/UtilsApp.java
@@ -1,8 +1,11 @@
 package com.blankj.launcher.app;
 
 import android.content.Context;
+import android.graphics.Bitmap;
+import android.graphics.Canvas;
+import android.view.View;
 
-import com.blankj.base.BaseApplication;
+import com.blankj.lib.base.BaseApplication;
 
 /**
  * <pre>
@@ -29,6 +32,20 @@ public class UtilsApp extends BaseApplication {
     public void onCreate() {
         super.onCreate();
         sInstance = this;
+    }
+
+
+    private Bitmap getBitmapForView(View src) {
+        Bitmap bitmap = Bitmap.createBitmap(
+                src.getWidth(),
+                src.getHeight(),
+                Bitmap.Config.ARGB_8888
+        );
+
+        Canvas canvas = new Canvas(bitmap);
+        src.draw(canvas);
+
+        return bitmap;
     }
 }