+++ lib/base/src/main/java/com/blankj/lib/base/rv/BaseViewHolder.java
@@ -1,4 +1,4 @@
-package com.blankj.base.rv;
+package com.blankj.lib.base.rv;
 
 import android.support.annotation.IdRes;
 import android.support.v7.widget.RecyclerView;