+++ utilcode/pkg/src/main/java/com/blankj/utilcode/pkg/feature/CoreUtilActivity.java
@@ -6,11 +6,13 @@ import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.view.View;
 
-import com.blankj.base.BaseBackActivity;
-import com.blankj.utilcode.R;
+import com.blankj.lib.base.BaseBackActivity;
+import com.blankj.utilcode.pkg.R;
 import com.blankj.utilcode.pkg.feature.activity.ActivityActivity;
+import com.blankj.utilcode.pkg.feature.adaptScreen.AdaptScreenActivity;
 import com.blankj.utilcode.pkg.feature.app.AppActivity;
 import com.blankj.utilcode.pkg.feature.bar.BarActivity;
+import com.blankj.utilcode.pkg.feature.blur.BlurActivity;
 import com.blankj.utilcode.pkg.feature.clean.CleanActivity;
 import com.blankj.utilcode.pkg.feature.device.DeviceActivity;
 import com.blankj.utilcode.pkg.feature.fragment.FragmentActivity;
@@ -31,7 +33,6 @@ import com.blankj.utilcode.pkg.feature.sp.SPActivity;
 import com.blankj.utilcode.pkg.feature.span.SpanActivity;
 import com.blankj.utilcode.pkg.feature.toast.ToastActivity;
 import com.blankj.utilcode.util.BusUtils;
-import com.blankj.utilcode.util.Utils;
 
 /**
  * <pre>
@@ -61,7 +62,6 @@ public class CoreUtilActivity extends BaseBackActivity {
 
     @Override
     public void initView(Bundle savedInstanceState, View contentView) {
-        Utils.getApp();
         setTitle(R.string.core_util);
     }
 
@@ -84,7 +84,7 @@ public class CoreUtilActivity extends BaseBackActivity {
     }
 
     public void adaptScreenClick(View view) {
-        BusUtils.post("AdaptScreenActivity#start", this);
+        AdaptScreenActivity.Companion.start(this);
     }
 
     public void appClick(View view) {
@@ -93,6 +93,10 @@ public class CoreUtilActivity extends BaseBackActivity {
 
     public void barClick(View view) {
         BarActivity.start(this);
+    }
+
+    public void blurClick(View view) {
+        BlurActivity.start(this);
     }
 
     public void cleanClick(View view) {