+++ utilcode/pkg/src/main/java/com/blankj/utilcode/pkg/feature/adaptScreen/WidthActivity.kt
@@ -7,7 +7,7 @@ import android.graphics.Color
 import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
 import android.view.WindowManager
-import com.blankj.utilcode.R
+import com.blankj.utilcode.pkg.R
 import com.blankj.utilcode.util.AdaptScreenUtils
 import kotlinx.android.synthetic.main.activity_adapt_width.*