+++ utilcode/pkg/src/main/java/com/blankj/utilcode/pkg/feature/app/AppActivity.java
@@ -7,9 +7,9 @@ import android.support.annotation.Nullable;
 import android.view.View;
 import android.widget.TextView;
 
-import com.blankj.base.BaseBackActivity;
+import com.blankj.lib.base.BaseBackActivity;
 import com.blankj.utilcode.pkg.Config;
-import com.blankj.utilcode.R;
+import com.blankj.utilcode.pkg.R;
 import com.blankj.utilcode.pkg.helper.PermissionHelper;
 import com.blankj.utilcode.util.AppUtils;
 import com.blankj.utilcode.util.FileUtils;