+++ utilcode/pkg/src/main/java/com/blankj/utilcode/pkg/feature/bar/BarStatusAlphaFragment.java
@@ -6,8 +6,8 @@ import android.view.View;
 import android.widget.SeekBar;
 import android.widget.TextView;
 
-import com.blankj.base.BaseLazyFragment;
-import com.blankj.utilcode.R;
+import com.blankj.lib.base.BaseLazyFragment;
+import com.blankj.utilcode.pkg.R;
 import com.blankj.utilcode.util.BarUtils;
 import com.blankj.utilcode.util.LogUtils;