+++ utilcode/pkg/src/main/java/com/blankj/utilcode/pkg/feature/bar/BarStatusFragmentActivity.java
@@ -12,8 +12,8 @@ import android.support.v4.view.ViewPager;
 import android.view.MenuItem;
 import android.view.View;
 
-import com.blankj.utilcode.R;
-import com.blankj.base.BaseActivity;
+import com.blankj.utilcode.pkg.R;
+import com.blankj.lib.base.BaseActivity;
 
 import java.util.ArrayList;