+++ utilcode/pkg/src/main/java/com/blankj/utilcode/pkg/feature/path/PathActivity.java
@@ -7,8 +7,8 @@ import android.support.annotation.Nullable;
 import android.view.View;
 import android.widget.TextView;
 
-import com.blankj.utilcode.R;
-import com.blankj.base.BaseBackActivity;
+import com.blankj.utilcode.pkg.R;
+import com.blankj.lib.base.BaseBackActivity;
 import com.blankj.utilcode.util.PathUtils;
 import com.blankj.utilcode.util.SpanUtils;