+++ utilcode/pkg/src/main/java/com/blankj/utilcode/pkg/feature/phone/PhoneActivity.java
@@ -8,8 +8,8 @@ import android.support.annotation.Nullable;
 import android.view.View;
 import android.widget.TextView;
 
-import com.blankj.base.BaseBackActivity;
-import com.blankj.utilcode.R;
+import com.blankj.lib.base.BaseBackActivity;
+import com.blankj.utilcode.pkg.R;
 import com.blankj.utilcode.pkg.helper.PermissionHelper;
 import com.blankj.utilcode.util.PhoneUtils;
 import com.blankj.utilcode.util.SpanUtils;